split out forum post code into its own resource
