+++ server/src/resources/forum/type.go
@@ -1,83 +1,1 @@
 package forum
-
-import (
-	"time"
-
-	"github.com/openmultiplayer/web/server/src/db"
-	"github.com/openmultiplayer/web/server/src/resources/forum/category"
-)
-
-type Author struct {
-	ID        string    `json:"id"`
-	Name      string    `json:"name"`
-	Admin     bool      `json:"admin"`
-	CreatedAt time.Time `json:"createdAt"`
-}
-
-type Post struct {
-	ID         string             `json:"id"`
-	Title      *string            `json:"title"`
-	Slug       *string            `json:"slug"`
-	Body       string             `json:"body"`
-	Short      string             `json:"short"`
-	First      bool               `json:"first"`
-	CreatedAt  time.Time          `json:"createdAt"`
-	UpdatedAt  time.Time          `json:"updatedAt"`
-	DeletedAt  *time.Time         `json:"deletedAt"`
-	UserID     string             `json:"userId"`
-	RootPostID *string            `json:"rootPostId"`
-	Author     Author             `json:"author"`
-	Tags       []string           `json:"tags"`
-	Posts      int                `json:"posts"`
-	ReplyTo    *category.PostMeta `json:"replyTo"`
-	Category   category.Category  `json:"category"`
-}
-
-type Tag struct {
-	ID    string `json:"id"`
-	Name  string `json:"name"`
-	Posts int    `json:"posts"`
-}
-
-func FromModel(u *db.PostModel) (w *Post) {
-	tags := []string{}
-	for _, t := range u.RelationsPost.Tags {
-		tags = append(tags, t.Name)
-	}
-
-	cat := category.Category{}
-	if u.RelationsPost.Category != nil {
-		cat = *category.FromModel(u.RelationsPost.Category)
-	}
-
-	var replyTo *category.PostMeta
-	if u.RelationsPost.ReplyTo != nil {
-		replyTo = &category.PostMeta{
-			Author: u.RelationsPost.Author.Name,
-			PostID: u.RelationsPost.ReplyTo.ID,
-		}
-	}
-
-	return &Post{
-		ID:         u.InnerPost.ID,
-		Slug:       u.InnerPost.Slug,
-		Title:      u.InnerPost.Title,
-		Body:       u.InnerPost.Body,
-		Short:      u.InnerPost.Short,
-		First:      u.InnerPost.First,
-		CreatedAt:  u.InnerPost.CreatedAt,
-		UpdatedAt:  u.InnerPost.UpdatedAt,
-		DeletedAt:  u.InnerPost.DeletedAt,
-		UserID:     u.InnerPost.UserID,
-		RootPostID: u.InnerPost.RootPostID,
-		Author: Author{
-			ID:        u.RelationsPost.Author.InnerUser.ID,
-			Name:      u.RelationsPost.Author.InnerUser.Name,
-			Admin:     u.RelationsPost.Author.InnerUser.Admin,
-			CreatedAt: u.RelationsPost.Author.InnerUser.CreatedAt,
-		},
-		Tags:     tags,
-		Category: cat,
-		ReplyTo:  replyTo,
-	}
-}
